requirements.txt
aiogram==3.7.0
mysql-connector=2.2.9